Update method to show a notification
