+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -1708,15 +1708,13 @@ public class ChatClient internal constructor(
         }
 
         @Throws(IllegalStateException::class)
-        internal suspend fun displayNotificationWithData(
-            channelType: String,
-            channelId: String,
-            messageId: String,
+        internal fun displayNotification(
+            channel: Channel,
+            message: Message,
         ) {
-            ensureClientInitialized().notifications.displayNotificationWithData(
-                channelId = channelId,
-                channelType = channelType,
-                messageId = messageId,
+            ensureClientInitialized().notifications.displayNotification(
+                channel = channel,
+                message = message,
             )
         }
